+++ WHATSNEW.md
@@ -10,6 +10,7 @@
  * fixed smugmug.com plugin
  + added pinterest.com pin-board widget plugin
  + added pinterest.com user widget plugin
+ + added pinterest.com pin widget plugin
 
 0.3.10